+++ src/Exception/InvalidParentException.php
@@ -1,0 +1,16 @@
+<?php
+
+/**
+ * (c) Dennis Meckel
+ *
+ * For the full copyright and license information,
+ * please view the LICENSE file that was distributed with this source code.
+ */
+
+namespace Rayne\wz2008\Graph\Exception;
+
+use InvalidArgumentException;
+
+class InvalidParentException extends InvalidArgumentException
+{
+}